fix(runtime): address remote matcher review gaps
